requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Werkzeug==3.1.3
Jinja2==3.1.2
WTForms==3.2.1
SQLAlchemy==2.0.43
itsdangerous==2.2.0
click==8.2.1
MarkupSafe==2.1.2
openai==1.3.0
stripe==7.8.0
requests==2.31.0
python-dotenv==1.0.0
uuid==1.30

# AI Model Dependencies (Optional - for enhanced analysis)
transformers>=4.20.0
torch>=1.12.0
sentence-transformers>=2.2.0
numpy>=1.21.0
scikit-learn>=1.0.0
